+++ CHANGES.rst
@@ -5,8 +5,8 @@ Version 2.3.4
 
 Unreleased
 
--   ``Authorization.from_header`` detects tokens that end with base64 padding (``=``).
-    :issue:`2685`
+-   ``Authorization.from_header`` and ``WWWAuthenticate.from_header`` detects tokens
+    that end with base64 padding (``=``). :issue:`2685`
 
 
 Version 2.3.3